BibHarvest: various fixes for Pos
* Already existing records are only appended
  with the fulltext and an external link to it.

* Adds a summary at the end of the bibched log.

* Sends a summary via email to CFG_POSHARVESTEMAIL.

* Renames CFG_FULLTEXT_DOWNLOAD_DIR variable to
  CFG_POS_OUT_DIRECTORY.

* Keeps a copy of .append.xml, .insert.xml and
  .not_found.xml collections to CFG_POS_OUT_DIRECTORY.

* Updates bibfilter_pos_fulltext_uploader to use
  pos_package from harvestingkit.

* Renames the filter bibfilter_pos_fulltext_uploader.py
  to bibfilter_oaipos2inspire.py.

Signed-off by: Georgios Papoutsakis <georgios.papoutsakis@cern.ch>
